24/5/2019 update Buy and stock

+++ src/helper/UsefulCalculas.java
@@ -8,8 +8,8 @@ public class UsefulCalculas {
     
     ProductDAO productDAO = new ProductDAO();
 
-    public int getUnitsFromHoleWeight(float weight_all_units, int weight_of_unit) {
-        int result = 0;
+    public float getUnitsFromHoleWeight(float weight_all_units, float weight_of_unit) {
+        float result = 0;
         float weight_all_units_byGM = weight_all_units * 1000;
         result = (int) (weight_all_units_byGM / weight_of_unit);
         return result;
@@ -48,7 +48,7 @@ public class UsefulCalculas {
         int result = 0;
         try {
             float alertWeght = productDAO.getProductById(id).getAllertWeight();
-            int numberUnit = getUnitsFromHoleWeight(productDAO.getProductById(id).getUnitsWeightInStock(),
+            float numberUnit = getUnitsFromHoleWeight(productDAO.getProductById(id).getUnitsWeightInStock(),
                     productDAO.getProductById(id).getProductWeight());
             result = (int) (numberUnit - alertWeght);
         } catch (Exception e) {
